refactor (#32) : 한정판매 주문 생성 시 요청 할 feign client 관련 초기 세팅
