+++ src/main/java/com/limito/order/common/feignClient/LimitedFeignClient.java
@@ -7,7 +7,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import com.limito.order.cart.domain.dto.feignClient.GetPurchaseAmountLimitRequestV1;
 import com.limito.order.cart.domain.dto.feignClient.GetPurchaseAmountLimitResponseV1;
-import com.limito.order.cart.domain.dto.feignClient.ReserveStockRequestV1;
+import com.limito.order.order.domain.dto.feignClient.limited.ReduceStockRequestV1;
+import com.limito.order.order.domain.dto.feignClient.limited.ReserveStockRequestV1;
 
 import jakarta.validation.Valid;
 
@@ -21,4 +22,7 @@ public interface LimitedFeignClient {
 
 	@PostMapping("/internal/v1/limited-products/stock/reserve")
 	ResponseEntity<Void> reserveStock(@Valid @RequestBody ReserveStockRequestV1 request);
+
+	@PostMapping("/stock/reduce")
+	ResponseEntity<Void> reduceStock(@Valid @RequestBody ReduceStockRequestV1 request);
 }
