+++ src/main/java/com/limito/order/order/domain/dto/feignClient/limited/ReduceStockProductRequest.java
@@ -1,0 +1,18 @@
+package com.limito.order.order.domain.dto.feignClient.limited;
+
+import java.util.UUID;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+
+public class ReduceStockProductRequest {
+	@NotNull(message = "옵션 id는 null일 수 없습니다.")
+	UUID limitedProductOptionId;
+
+	@NotNull(message = "아이템 id는 null일 수 없습니다.")
+	UUID limitedProductItemId;
+
+	@NotNull(message = "수량은 null일 수 없습니다.")
+	@Positive(message = "최소 구매 수량은 1개입니다.")
+	int amount;
+}
